backend/requirements-modified.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
websockets==11.0.3

# API communication
httpx==0.25.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
pytest-cov==4.1.0

# Development tools
black==23.10.1
isort==5.12.0
flake8==6.1.0